+++ CHANGELOG.md
@@ -6,6 +6,11 @@
   reexport from Nix.
   (#[21](https://github.com/asomers/mio-aio/pull/21))
 
+- Updated mio to 0.7.
+
+- Added a `tokio` feature flag, which enables extra methods needed by a mio-aio
+  consumers that wish to implement Tokio's `AioSource` trait.
+
 ## [0.5.0] - 2021-05-31
 
 ### Changed